Fix for whereFlags not being available in the filterbuilder because it was moved from the HasFilter trait to the QueryBuilder.
