+++ src/Concerns/HasFilter.php
@@ -5,6 +5,7 @@ namespace Level23\Druid\Concerns;
 
 use Closure;
 use InvalidArgumentException;
+use Level23\Druid\Types\DataType;
 use Level23\Druid\Filters\InFilter;
 use Level23\Druid\Filters\OrFilter;
 use Level23\Druid\Filters\AndFilter;
@@ -23,6 +24,7 @@ use Level23\Druid\Filters\FilterInterface;
 use Level23\Druid\Filters\JavascriptFilter;
 use Level23\Druid\Interval\IntervalInterface;
 use Level23\Druid\Dimensions\DimensionBuilder;
+use Level23\Druid\VirtualColumns\VirtualColumn;
 use Level23\Druid\Extractions\ExtractionBuilder;
 use Level23\Druid\Dimensions\DimensionInterface;
 use Level23\Druid\Filters\ColumnComparisonFilter;
@@ -34,6 +36,24 @@ trait HasFilter
      * @var \Level23\Druid\Filters\FilterInterface|null
      */
     protected $filter;
+
+    /**
+     * @var array|\Level23\Druid\VirtualColumns\VirtualColumnInterface[]
+     */
+    protected $virtualColumns = [];
+
+    /**
+     * This contains a list of "temporary" field names which we will use to store our result of
+     * a virtual column when the whereFlag() method is used.
+     *
+     * @var array
+     */
+    protected $placeholders = [];
+
+    /**
+     * @var \Level23\Druid\DruidClient
+     */
+    protected $client;
 
     /**
      * Filter our results where the given dimension matches the value based on the operator.
@@ -626,6 +646,87 @@ trait HasFilter
     }
 
     /**
+     * Filter on records which match using a bitwise AND comparison.
+     *
+     * Only records will match where the dimension contains ALL bits which are also enabled in the given $flags
+     * argument. Support for 64 bit integers are supported.
+     *
+     * Druid has support for bitwise flags since version 0.20.2.
+     * Before that, we have build our own variant, but then javascript support is required.
+     *
+     * JavaScript-based functionality is disabled by default. Please refer to the Druid JavaScript programming guide
+     * for guidelines about using Druid's JavaScript functionality, including instructions on how to enable it:
+     * https://druid.apache.org/docs/latest/development/javascript.html
+     *
+     * @param string $dimension The dimension which contains int values where you want to do a bitwise AND check
+     *                          against.
+     * @param int    $flags     The bit's which you want to check if they are enabled in the given dimension.
+     *
+     * @return $this
+     */
+    public function orWhereFlags(string $dimension, int $flags)
+    {
+        return $this->whereFlags($dimension, $flags, 'or');
+    }
+
+    /**
+     * Filter on records which match using a bitwise AND comparison.
+     *
+     * Only records will match where the dimension contains ALL bits which are also enabled in the given $flags
+     * argument. Support for 64 bit integers are supported.
+     *
+     * Druid has support for bitwise flags since version 0.20.2.
+     * Before that, we have build our own variant, but then javascript support is required.
+     *
+     * JavaScript-based functionality is disabled by default. Please refer to the Druid JavaScript programming guide
+     * for guidelines about using Druid's JavaScript functionality, including instructions on how to enable it:
+     * https://druid.apache.org/docs/latest/development/javascript.html
+     *
+     * @param string $dimension The dimension which contains int values where you want to do a bitwise AND check
+     *                          against.
+     * @param int    $flags     The bit's which you want to check if they are enabled in the given dimension.
+     * @param string $boolean   This influences how this filter will be joined with previous added filters. Should both
+     *                          filters apply ("and") or one or the other ("or") ? Default is "and".
+     *
+     * @return $this
+     */
+    public function whereFlags(string $dimension, int $flags, string $boolean = 'and')
+    {
+        /**
+         * If our version supports this, let's use the new and improved bitwiseAnd function!
+         */
+        $version = (string)$this->client->config('version');
+        if (!empty($version) && version_compare($version, '0.20.2', '>=')) {
+
+            $placeholder          = 'v' . count($this->placeholders);
+            $this->placeholders[] = $placeholder;
+
+            $this->virtualColumn('bitwiseAnd("' . $dimension . '", ' . $flags . ')', $placeholder, DataType::LONG);
+
+            return $this->where($placeholder, '=', $flags, null, $boolean);
+        }
+
+        return $this->where($dimension, '=', $flags, function (ExtractionBuilder $extraction) use ($flags) {
+            // Do a binary "AND" flag comparison on a 64 bit int. The result will either be the
+            // $flags, or 0 when it's bit is not set.
+            $extraction->javascript('
+                function(dimensionValue) { 
+                    var givenValue = ' . $flags . '; 
+                    var hi = 0x80000000; 
+                    var low = 0x7fffffff; 
+                    var hi1 = ~~(dimensionValue / hi); 
+                    var hi2 = ~~(givenValue / hi); 
+                    var low1 = dimensionValue & low; 
+                    var low2 = givenValue & low; 
+                    var h = hi1 & hi2; 
+                    var l = low1 & low2; 
+                    return (h*hi + l); 
+                }
+            ');
+        }, $boolean);
+    }
+
+    /**
      * Filter on an dimension where the value exists in the given intervals array.
      *
      * The intervals array can contain the following:
@@ -826,5 +927,32 @@ trait HasFilter
         call_user_func($extraction, $builder);
 
         return $builder->getExtraction();
+    }
+
+    /**
+     * Create a virtual column.
+     *
+     * Virtual columns are queryable column "views" created from a set of columns during a query.
+     *
+     * A virtual column can potentially draw from multiple underlying columns, although a virtual column always
+     * presents itself as a single column.
+     *
+     * Virtual columns can be used as dimensions or as inputs to aggregators.
+     *
+     * NOTE: virtual columns are NOT automatically added to your output. You should select it separately if you want to
+     * add it also to your output. Use selectVirtual to do both at once.
+     *
+     * @param string $expression
+     * @param string $as
+     * @param string $outputType
+     *
+     * @return $this
+     * @see https://druid.apache.org/docs/latest/misc/math-expr.html
+     */
+    public function virtualColumn(string $expression, string $as, $outputType = DataType::STRING)
+    {
+        $this->virtualColumns[] = new VirtualColumn($expression, $as, $outputType);
+
+        return $this;
     }
 }
